# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velo API",
    "version": "0.1.0",
    "description": "Velo — anonymous cash liquidity on Stellar. A lightweight, payment-aware service interface for agent-assisted and application-driven cash flows. Paid routes use an x402-style challenge: call without payment to receive a challenge, pay the quoted USDC amount on Stellar with the `velo:request` memo, then retry with the transaction hash in the `X-Payment` header. Each payment transaction is single-use.",
    "contact": {
      "name": "Nullifier Systems",
      "url": "https://github.com/Nullifier-Systems/velo"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/Nullifier-Systems/velo/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "tags": [
    {
      "name": "meta",
      "description": "Health, discovery, and specification endpoints."
    },
    {
      "name": "cash",
      "description": "Cash request lifecycle: discover providers, lock escrow, poll, release."
    },
    {
      "name": "reputation",
      "description": "On-chain reputation lookups."
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "tags": [
          "meta"
        ],
        "summary": "Health check",
        "description": "Free liveness probe for infrastructure.",
        "x-rate-limit": {
          "max": 100,
          "timeWindow": "1 minute"
        },
        "responses": {
          "200": {
            "description": "Service is up.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "getOpenApiSpec",
        "tags": [
          "meta"
        ],
        "summary": "OpenAPI specification",
        "description": "Free machine-readable OpenAPI 3.1 document describing every endpoint, request/response shape, and x402 price of this API.",
        "x-rate-limit": {
          "max": 60,
          "timeWindow": "1 minute"
        },
        "responses": {
          "200": {
            "description": "The OpenAPI document itself.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "OpenAPI 3.1 document."
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/services": {
      "get": {
        "operationId": "listServices",
        "tags": [
          "meta"
        ],
        "summary": "Service catalog",
        "description": "Free catalog of paid endpoints and their x402 prices, for agent autodiscovery.",
        "x-rate-limit": {
          "max": 60,
          "timeWindow": "1 minute"
        },
        "responses": {
          "200": {
            "description": "Catalog of paid services.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "services"
                  ],
                  "properties": {
                    "services": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "endpoint",
                          "price_usdc"
                        ],
                        "properties": {
                          "endpoint": {
                            "type": "string",
                            "description": "Method and path of the paid endpoint.",
                            "examples": [
                              "POST /api/v1/cash/request"
                            ]
                          },
                          "price_usdc": {
                            "type": "string",
                            "description": "Price per call in USDC, as a decimal string.",
                            "examples": [
                              "0.01"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/cash/agents": {
      "get": {
        "operationId": "findCashAgents",
        "tags": [
          "cash"
        ],
        "summary": "Provider discovery",
        "description": "Find nearby cash providers. Paid route: 0.001 USDC per call via x402.",
        "security": [
          {
            "x402Payment": []
          }
        ],
        "x-price-usdc": "0.001",
        "x-rate-limit": {
          "max": 30,
          "timeWindow": "1 minute"
        },
        "responses": {
          "200": {
            "description": "Nearby cash providers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "agents"
                  ],
                  "properties": {
                    "agents": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CashAgent"
                      }
                    }
                  }
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/cash/request": {
      "post": {
        "operationId": "createCashRequest",
        "tags": [
          "cash"
        ],
        "summary": "Create a cash request",
        "description": "Lock funds in the escrow contract and return a claim URL + QR payload for the cash hand-off. The redemption secret is held client-side and never returned by the API. Paid route: 0.01 USDC per call via x402.",
        "security": [
          {
            "x402Payment": []
          }
        ],
        "x-price-usdc": "0.01",
        "x-rate-limit": {
          "max": 20,
          "timeWindow": "1 minute"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Escrow locked; cash request created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRequestCreated"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "description": "The escrow lock transaction failed on-chain.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cash/request/{id}": {
      "get": {
        "operationId": "getCashRequest",
        "tags": [
          "cash"
        ],
        "summary": "Poll request status",
        "description": "Free polling endpoint for a pending cash request. The secret is never included.",
        "x-rate-limit": {
          "max": 60,
          "timeWindow": "1 minute"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Trade id returned when the request was created (64-char hex).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current state of the cash request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRequestStatus"
                }
              }
            }
          },
          "404": {
            "description": "No cash request with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/v1/cash/request/{id}/release": {
      "post": {
        "operationId": "releaseCashRequest",
        "tags": [
          "cash"
        ],
        "summary": "Release escrow (hand-off)",
        "description": "Merchant confirms the cash hand-off and releases the escrow using the secret embedded in the scanned QR. Free — this is a state-transition call, not a discovery call.",
        "x-rate-limit": {
          "max": 20,
          "timeWindow": "1 minute"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Trade id of the locked cash request.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "secret"
                ],
                "properties": {
                  "secret": {
                    "type": "string",
                    "description": "Redemption secret from the scanned QR (hex preimage of secret_hash)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Escrow released.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "status"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "const": "released"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing secret.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No cash request with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Request is not in the `locked` state (already released or refunded).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "description": "The escrow release transaction failed on-chain.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reputation/{address}": {
      "get": {
        "operationId": "getReputation",
        "tags": [
          "reputation"
        ],
        "summary": "On-chain reputation lookup",
        "description": "Trust signal for a Stellar address. Paid route: 0.0005 USDC per call via x402.",
        "security": [
          {
            "x402Payment": []
          }
        ],
        "x-price-usdc": "0.0005",
        "x-rate-limit": {
          "max": 30,
          "timeWindow": "1 minute"
        },
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Stellar account address (G...).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reputation summary. Fields are null until the on-chain reputation source is wired up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reputation"
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Payment",
        "description": "x402 payment header: the hash of a successful Stellar transaction paying the merchant at least the operation's `x-price-usdc` with memo `velo:request`. Each transaction hash is accepted once. Requests without a valid payment receive 402 with a challenge, not 401."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "detail": {
            "type": "string",
            "description": "Optional underlying error detail."
          }
        }
      },
      "PaymentChallenge": {
        "type": "object",
        "description": "x402 challenge returned when no X-Payment header is present. Pay `amount_usdc` to `pay_to` with the given memo, then retry the request with the transaction hash in the X-Payment header.",
        "required": [
          "challenge"
        ],
        "properties": {
          "challenge": {
            "type": "object",
            "required": [
              "amount_usdc",
              "pay_to",
              "memo"
            ],
            "properties": {
              "amount_usdc": {
                "type": "string",
                "description": "Price of this call in USDC, as a decimal string.",
                "examples": [
                  "0.01"
                ]
              },
              "pay_to": {
                "type": "string",
                "description": "Stellar address (G...) of the merchant to pay."
              },
              "memo": {
                "type": "string",
                "description": "Memo that MUST be attached to the payment transaction.",
                "examples": [
                  "velo:request"
                ]
              }
            }
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "const": 429
          },
          "error": {
            "type": "string",
            "const": "Too Many Requests"
          },
          "message": {
            "type": "string"
          },
          "retryAfter": {
            "type": "string",
            "description": "Human-readable wait time, e.g. \"1 minute\"."
          },
          "retryAfterSeconds": {
            "type": "integer",
            "description": "Seconds to wait before retrying."
          }
        }
      },
      "CashAgent": {
        "type": "object",
        "required": [
          "name",
          "distance_km",
          "tier"
        ],
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "Farmacia Guadalupe"
            ]
          },
          "distance_km": {
            "type": "number",
            "examples": [
              0.3
            ]
          },
          "tier": {
            "type": "string",
            "examples": [
              "Maestro"
            ]
          }
        }
      },
      "CashRequestBody": {
        "type": "object",
        "required": [
          "seller",
          "buyer",
          "amount_stroops",
          "secret_hash"
        ],
        "properties": {
          "seller": {
            "type": "string",
            "description": "Stellar address (G...) of the cash provider."
          },
          "buyer": {
            "type": "string",
            "description": "Stellar address (G...) of the person requesting cash."
          },
          "amount_stroops": {
            "type": "string",
            "description": "Amount in stroops as a bigint string, e.g. \"10000000\" = 1 unit.",
            "examples": [
              "10000000"
            ]
          },
          "secret_hash": {
            "type": "string",
            "description": "64-character hex SHA-256 hash of the client-held redemption secret.",
            "pattern": "^[0-9a-fA-F]{64}$"
          }
        }
      },
      "CashRequestCreated": {
        "type": "object",
        "required": [
          "claim_url",
          "qr_payload",
          "instructions"
        ],
        "properties": {
          "claim_url": {
            "type": "string",
            "description": "URL the buyer opens/shares to claim the cash.",
            "examples": [
              "https://app.velo.cash/claim/{trade_id}"
            ]
          },
          "qr_payload": {
            "type": "string",
            "description": "Deep-link payload to encode as a QR code.",
            "examples": [
              "velo://claim?request_id={trade_id}&contract={contract_id}"
            ]
          },
          "instructions": {
            "type": "string"
          }
        }
      },
      "CashRequestStatus": {
        "type": "object",
        "description": "Public view of a cash request. The redemption secret is never included.",
        "required": [
          "id",
          "contractId",
          "seller",
          "buyer",
          "amountStroops",
          "secretHashHex",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Trade id (64-char hex)."
          },
          "contractId": {
            "type": "string",
            "description": "Escrow contract id (C...)."
          },
          "seller": {
            "type": "string"
          },
          "buyer": {
            "type": "string"
          },
          "amountStroops": {
            "type": "string"
          },
          "secretHashHex": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "locked",
              "released",
              "refunded"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Reputation": {
        "type": "object",
        "required": [
          "address",
          "completion_rate",
          "trades",
          "trusted"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "completion_rate": {
            "type": [
              "number",
              "null"
            ]
          },
          "trades": {
            "type": [
              "integer",
              "null"
            ]
          },
          "trusted": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required (x402). Returned when the X-Payment header is missing (body contains a challenge), already used, or does not reference a successful Stellar transaction paying the merchant the required amount with the `velo:request` memo.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/PaymentChallenge"
                },
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded for this route (per client IP).",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RateLimitError"
            }
          }
        }
      }
    }
  }
}